% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MC_Trend.R
\name{MC_Trend}
\alias{MC_Trend}
\title{Monte Carlo Trend Analysis}
\usage{
MC_Trend(x, n_rep, plot_title)
}
\arguments{
\item{x}{A data frame containing the input data. The first raw expected to contain model names or time series names.}

\item{n_rep}{Number of replications for the Monte Carlo simulation.}

\item{plot_title}{Title for the plot.}
}
\value{
A data frame and a plot containing results of the trend analysis.
}
\description{
This function performs Monte Carlo trend analysis on input data and generates plots.
}
