% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCTrend.R
\name{MCTrend}
\alias{MCTrend}
\title{Monte Carlo Trend Analysis}
\usage{
MCTrend(x, n_rep, plot_title, opt)
}
\arguments{
\item{x}{A data frame containing the input data. The first raw expected to contain model names or time series names.}

\item{n_rep}{Number of replications for the Monte Carlo simulation.}

\item{plot_title}{Title for the plot.}

\item{opt}{A number indicating type of results, for opt = 1 returns test result, opt = 2 returns plot}
}
\value{
A data frame and a plot containing results of the trend analysis.
}
\description{
This function performs Monte Carlo trend analysis on input data and generates plots.
}
